requirements.txt
streamlit
pandas
google-cloud-vision
google-auth
gspread
opencv-python-headless
thefuzz
numpy